requirements.txt
haversine==2.9.0
plotly==6.0.0
matplotlib==3.9.3
seaborn==0.13.2
pandas==2.2.2
numpy==1.26.4
streamlit==1.41.1
pillow==11.0.0
